+++ src/init.c
@@ -32,6 +32,7 @@ SEXP nano_ListenerSymbol;
 SEXP nano_PipeSymbol;
 SEXP nano_ProtocolSymbol;
 SEXP nano_RawSymbol;
+SEXP nano_RefhookSymbol;
 SEXP nano_ResponseSymbol;
 SEXP nano_ResultSymbol;
 SEXP nano_RtcSymbol;
@@ -76,6 +77,7 @@ static void RegisterSymbols(void) {
   nano_PipeSymbol = Rf_install("pipe");
   nano_ProtocolSymbol = Rf_install("protocol");
   nano_RawSymbol = Rf_install("raw");
+  nano_RefhookSymbol = Rf_install("refhook");
   nano_ResponseSymbol = Rf_install("response");
   nano_ResultSymbol = Rf_install("result");
   nano_RtcSymbol = Rf_install("rawToChar");
@@ -151,7 +153,7 @@ static const R_CallMethodDef callMethods[] = {
   {"rnng_ctx_create", (DL_FUNC) &rnng_ctx_create, 1},
   {"rnng_ctx_open", (DL_FUNC) &rnng_ctx_open, 1},
   {"rnng_cv_alloc", (DL_FUNC) &rnng_cv_alloc, 0},
-  {"rnng_cv_recv_aio", (DL_FUNC) &rnng_cv_recv_aio, 7},
+  {"rnng_cv_recv_aio", (DL_FUNC) &rnng_cv_recv_aio, 8},
   {"rnng_cv_request", (DL_FUNC) &rnng_cv_request, 8},
   {"rnng_cv_reset", (DL_FUNC) &rnng_cv_reset, 1},
   {"rnng_cv_signal", (DL_FUNC) &rnng_cv_signal, 1},
@@ -180,10 +182,10 @@ static const R_CallMethodDef callMethods[] = {
   {"rnng_protocol_open", (DL_FUNC) &rnng_protocol_open, 2},
   {"rnng_random", (DL_FUNC) &rnng_random, 1},
   {"rnng_recv", (DL_FUNC) &rnng_recv, 6},
-  {"rnng_recv_aio", (DL_FUNC) &rnng_recv_aio, 6},
+  {"rnng_recv_aio", (DL_FUNC) &rnng_recv_aio, 7},
   {"rnng_request", (DL_FUNC) &rnng_request, 7},
   {"rnng_send", (DL_FUNC) &rnng_send, 5},
-  {"rnng_send_aio", (DL_FUNC) &rnng_send_aio, 5},
+  {"rnng_send_aio", (DL_FUNC) &rnng_send_aio, 6},
   {"rnng_set_opt", (DL_FUNC) &rnng_set_opt, 3},
   {"rnng_sha1", (DL_FUNC) &rnng_sha1, 3},
   {"rnng_sha224", (DL_FUNC) &rnng_sha224, 3},